feat: updates template system with class name integration
Now uses dashes instead of periods. This allows it to use the same token for a class name applied to an element.
If it's found that the notice data is nullish then it attempts to add "display: none" style to any elements in the notice with a matching class name.