fix: adds styles for disabled buttons
There was no built in style for disabled buttons in the theme I was testing on, leading to a confusing user experience. Fixed the issue by reducing their opacity and preventing cursor interaction.

Also added the container element to all rule selectors to prevent them interfering with the main site's rules.